Move more early buffered lints to dyn lint diagnostics (3/N)
